Add functional login and create api key methods.
This commit adds a working method to authenticate to frost.po.et and get
a master api key good for 24 hours that will enable creation of 2nd
tier api keys that can create content only.  It also adds a method to
create those 2nd tier api keys.  This may be changed in the
near future, so I'm leaving this in the dev branch for now.